requirements.txt
Flask==3.0.2
requests==2.31.0  
pyodbc==5.1.0
azure-storage-blob==12.19.0
markdown==3.5.2
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-speech==1.36.0
openai==1.20.0